+++ packages/ir-sdk/fern/apis/ir-types-latest/changelog/CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [v64.0.0] - 2026-02-03
-- Feature: Add new members to `CursorPagination.page` to be consumed as the full URI, or full path (HATEOS-style)
+- Feature: Add new pagination types where the cursor is to be consumed as the full URI, or full path (HATEOS-style)
 
 ## [v63.2.0] - 2026-01-13
 - Feature: Add support for min and max validation keywords: